requirements/requirements.txt
django==3.1.6
django-debug-toolbar